Pandora: Bump to 1.8.6 * lazy: add modes back * lazy: tune proper userpaces tunables only * lazy: remove non-universal tunables * lazy: select blu_schedutil/schedutil/interactive as governors only * lazy: places block layer tunables to lazy main script instead of per mode changing * lazy: Disable RT_RUNTIME_SHARE and enable SCHEDTUNE_BOOST_UTIL * lazy: disable adreno idler via userspace * swift: stop touching crtc_commit threads via userspace * swift: move surfaceflinger to topapp stune for top-app boost benefits * swfft: script cleanup * swift: add ionice back * kcal: add deep black and white kcal preset Signed-off by: Alex Finhart <alexfinhart@gmail.com>

+++ customize.sh
@@ -75,17 +75,21 @@ ui_print ""
 ui_print " Setting executable permissions..."
 
 ui_print ""
-ui_print " [Notice]: toast.apk is an app for lazy's startup notification, if you feel unsecure about it, just install toast app manually"
+ui_print " [Notice]: please manually uninstall
+           old lazy app if an error pops-up or magisk can't install the app"
 ui_print ""
 
-# Do install-time script execution
-ui_print " Running Pandora on Boot..."
-stop pandora
-sh /data/adb/modules_update/toolbox8/system/bin/pandora
-
-# install notification app
+if [[ -f "/data/adb/modules_update/toolbox8/scripts/lazy" ]]
+then
+ui_print "Installing Lazy tweaks app.."
+pm install $MODPATH/base.apk
+ui_print "Installing Lazy notification app.."
 pm install $MODPATH/toast.apk
-
+fi
+# Do install-time script execution
+ui_print ""
+ui_print " Cleaning Set-Up"
+rm -rf /data/dalvik-cache
 ui_print ""
 ui_print "   Thank you for at least trying lazy tweaks! "
 ui_print "  		   Have a Great Day!"
